SPORT-45 Update to have values as double instead of int

+++ App/Workout/Models/ExerciseSet.cs
@@ -5,7 +5,7 @@ namespace SaveApp.App.Workout.Models
     public class ExerciseSet
     {
         public int Id { get; set; }
-        public int? Weight { get; set; }
+        public double? Weight { get; set; }
         public int? Reps { get; set; }
         public String? Notes { get; set; }
         public ExerciseSetPreviousValues? ExerciseSetPreviousValues { get; set; }
