+++ App/Workout/Repositories/Entities/WorkoutEntity.cs
@@ -4,7 +4,7 @@ namespace SaveApp.App.Workout.Repositories.Entities
     {
         public int Id { get; set; }
         public String? Name { get; set; }
-        public int? BodyWeight { get; set; }
+        public double? BodyWeight { get; set; }
         public DateTime Date { get; set; }
         public DateTime? StartTime { get; set; }
         public DateTime? EndTime { get; set; }
